할인 프로모션 파일 분리
